ASM_2

+++ assets/js/index.js
@@ -5,6 +5,28 @@ $("#add_user").submit(function(event){
     alert("Data Inserted Successfully!");
 })
 
+// $("#update_user").submit(function(event){
+//     event.preventDefault();
+//     var unindexed_array = $(this).serializeArray();
+//     var data = {}
+
+//     $.map(unindexed_array, function(n, i){
+//         data[n['name']] = n['value']
+//     })
+
+
+//     var request = {
+//         "url" : `http://localhost:8030/api/users/${data.id}`,
+//         "method" : "PUT",
+//         "data" : data
+//     }
+
+//     $.ajax(request).done(function(response){
+//         alert("Data Updated Successfully!");
+//     })
+
+// })
+
 $("#update_user").submit(function(event){
     event.preventDefault();
 
@@ -17,7 +39,7 @@ $("#update_user").submit(function(event){
 
 
     var request = {
-        "url" : `http://localhost:8080/api/users/${data.id}`,
+        "url" : `http://localhost:8030/api/users/${data.id}`,
         "method" : "PUT",
         "data" : data
     }
@@ -34,7 +56,52 @@ if(window.location.pathname == "/"){
         var id = $(this).attr("data-id")
 
         var request = {
-            "url" : `http://localhost:8080/api/users/${id}`,
+            "url" : `http://localhost:8030/api/users/${id}`,
+            "method" : "DELETE"
+        }
+
+        if(confirm("Do you really want to delete this record?")){
+            $.ajax(request).done(function(response){
+                alert("Data Deleted Successfully!");
+                location.reload();
+            })
+        }
+
+    })
+}
+
+
+
+///pro//////////////////////////
+
+
+$("#update_product").submit(function(event){
+    event.preventDefault();
+    var pro_array = $(this).serializeArray();
+    console.log(pro_array);
+    var data = {}
+    $.map(pro_array, function(n, i){
+        data[n['name']] = n['value']
+    })
+    var request = {
+        "url" : `http://localhost:8030/api/product/${data.id}`,
+        "method" : "PUT",
+        "data" : data
+    }
+
+    $.ajax(request).done(function(response){
+        alert("Data Updated Successfully!");
+    })
+
+})
+
+if(window.location.pathname == "/all-product"){
+    $ondelete = $(".table tbody td a.delete");
+    $ondelete.click(function(){
+        var id = $(this).attr("data-idpro")
+
+        var request = {
+            "url" : `http://localhost:8030/api/product/${id}`,
             "method" : "DELETE"
         }
